refactor onboarding pages + layouts
