+++ frontend/src/app/(password)/layout.tsx
@@ -1,0 +1,9 @@
+import OnboardingLayout from "@/components/onboarding/OnboardingLayout";
+
+export default function RootLayout({
+  children,
+}: Readonly<{
+  children: React.ReactNode;
+}>) {
+  return <OnboardingLayout>{children}</OnboardingLayout>;
+}
